fix : null値チェックを行うように修正しました。

+++ backend/src/main/java/jp/ac/dendai/spp/backend/service/LikeService.java
@@ -54,8 +54,11 @@ public class LikeService {
   public void deleteLike(UUID userId, LikeRequest request) {
     Like like = likeRepository.findByUserIdAndPostId(userId, request.getPostId());
 
-    likeRepository.delete(like);
+    if (like == null) {
+      throw new InvalidParameterException("いいねが見つかりません。");
+    }
 
+    likeRepository.delete(like);
     notificationService.deleteLikeNotification(like.getId());
   }
 }
